undo the aggressive refactoring (out of scope)
