+++ parallel-consumer-core/src/test-integration/java/io/confluent/parallelconsumer/integrationTests/RebalanceTest.java
@@ -4,33 +4,66 @@
  */
 package io.confluent.parallelconsumer.integrationTests;
 
+import io.confluent.parallelconsumer.ParallelConsumerOptions;
+import io.confluent.parallelconsumer.ParallelEoSStreamProcessor;
+import io.confluent.parallelconsumer.integrationTests.utils.KafkaClientUtils;
 import lombok.SneakyThrows;
+import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import pl.tlinkowski.unij.api.UniSets;
 
 import java.time.Duration;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
 
 import static io.confluent.parallelconsumer.ManagedTruth.assertThat;
+import static io.confluent.parallelconsumer.ParallelConsumerOptions.ProcessingOrder.PARTITION;
 import static io.confluent.parallelconsumer.integrationTests.utils.KafkaClientUtils.GroupOption.RESUE_GROUP;
 import static org.testcontainers.shaded.org.awaitility.Awaitility.await;
 import static org.testcontainers.shaded.org.hamcrest.Matchers.equalTo;
 import static org.testcontainers.shaded.org.hamcrest.Matchers.is;
 
 /**
- * todo docs
+ * Tests around what should happen when rebalancing occurs
  *
  * @author Antony Stubbs
  */
 class RebalanceTest extends BrokerIntegrationTest<String, String> {
 
+
+    Consumer<String, String> consumer;
+
+    ParallelEoSStreamProcessor<String, String> pc;
+
     public static final Duration INFINITE = Duration.ofDays(1);
 
+    {
+        super.numPartitions = 2;
+    }
+
+    @BeforeEach
+    void setup() {
+        setupTopic();
+        consumer = getKcu().createNewConsumer(KafkaClientUtils.GroupOption.NEW_GROUP);
+
+        pc = new ParallelEoSStreamProcessor<>(ParallelConsumerOptions.<String, String>builder()
+                .consumer(consumer)
+                .ordering(PARTITION) // just so we dont need to use keys
+                .build());
+
+        pc.subscribe(UniSets.of(topic));
+    }
+
+    /**
+     * Checks that when a rebalance happens, a final commit is done first for revoked partitions (that will be assigned
+     * to new consumers), so that the new consumer doesn't reprocess records that are already complete.
+     */
     @SneakyThrows
     @Test
     void commitUponRevoke() {
-        var recordsCount = 20;
+        var recordsCount = 20L;
         var count = new AtomicLong();
 
         //
@@ -41,7 +74,7 @@ class RebalanceTest extends BrokerIntegrationTest<String, String> {
 
         // consume all the messages
         pc.poll(recordContexts -> count.getAndIncrement());
-        await().untilAtomic(count, is(equalTo((long) recordsCount)));
+        await().untilAtomic(count, is(equalTo(recordsCount)));
 
         // cause rebalance
         var newConsumer = kcu.createNewConsumer(RESUE_GROUP);
